functionapp01/func/requirements.txt
azure-monitor-opentelemetry 
azure-functions
azure-identity
azure-ai-ml
openai